+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - rendering hints for query errors from EdgeDB.
 - support for EdgeDB binary protocol `2.0`.
+- new `EdgeDB.MultiRange` type to represent multiranges from `EdgeDB 4.0`.
 
 ## [0.6.1] - 2023-07-07